Fix missing tiles with GPU render and OIDN denoiser in Cycles X
Was noticed when testing upcoming tiled rendering with small tiles
sizes which stresses data transfers more than single tile rendering.

What happens is OIDN denoiser uses `copy_to_device()` using default
queue, which is not guaranteed to be synchronized before next call to
`copy_from_device()` happening via queue.

What we really want in the OIDN denoiser is synchronous data transfer
which is easiest to be achieved by creating a queue and synchronizing
it.

The issue is reported by Alaska in D12309#320253

Differential Revision: https://developer.blender.org/D12333