Make state tracking more reliable in Cycles X scheduler
Move logic to a re-usable function, so that state can be checked for
"regular" and "post-processing" render works.

Also corrected state tracking for tile written flag: it is a "sticky"
flag, meaning once the tile is written is always considered written.
Before this change it was possible that canceling render will write
tile again.

Should be no difference on user level, just preparing ground work for
tiles support.