Cycles X: Store passes and buffer params in tile file metadata
Allows to fully forget about passes and parameters once the file
was fully written.

Currently should be no functional changes. Will be used when a proper
memory saving is implemented: when EXR files are read from disk and
written to Blender after all view layers were rendered.

Is debatable why we need to store actual passes: for now it was the
most local change to move towards the goal (which is to decouple
full buffers reading from the Session/PathTrace/TileManager state).
Might end up storing simplified version of passes in the BufferParams
so that we can use `get_render_tile_pixels()` without having Scene
and only relying on the RenderBuffers. In any case, we would need to
store some sort of metadata, and whether it comes to/from passes is
something we can tweak in the future.

Differential Revision: https://developer.blender.org/D12447