Fix missing render progress on Windows in Cycles X
The issue is caused by `::wglShareLists()` returning an error if the
shared context is currently in use, so that the OpenGL context for
GPUDisplay did fail to be created.

After a lot of experimentation all the non-intrusive changes to support
OpenGL context creation with shared textures has failed.

The easiest way to have reliable OpenGL context creation is to do it on
the main thread.

For the offline rendering this is done in the pipeline.c. This change
takes advantage of this and uses the Render::gl_context for the GPU
display implementation for offline rendering.

For the viewport rendering we don't have Render, but the engine is
created from the main thread, so we can create context from the Cycles
side.

Not really happy with such approach, but this seems to be easiest and
safest way to go for now to get other bits of Cycles X finished. In the
future Vulkan should make things easier for the context creation and
multi-threaded applications.

Differential Revision: https://developer.blender.org/D12439